requirements.txt
google-api-python-client 
google-auth
python-dotenv
google_auth_oauthlib
google-generativeai